requirements.txt
paho-mqtt==1.5.0
rhasspy-hermes==0.2.1
webrtcvad==2.0.10
numpy>=1.18.0
